Created seeders and model factories. Modified existing migrations so that they work correctly.

+++ database/migrations/2020_07_07_173842_create_profiles_table.php
@@ -17,7 +17,7 @@ class CreateProfilesTable extends Migration
             $table->id();
             $table->timestamps();
 
-            $table->unsignedInteger('author_id')->unique();
+            $table->unsignedBigInteger('author_id')->unique();
             $table->foreign('author_id')->references('id')->on('authors');
         });
     }
